Create smaller modules by refactoring classes to separate modules
