+++ OTVision/domain/detection.py
@@ -1,0 +1,19 @@
+from dataclasses import dataclass
+from datetime import datetime
+
+
+@dataclass(frozen=True, repr=True)
+class Detection:
+    label: str
+    conf: float
+    x: float
+    y: float
+    w: float
+    h: float
+
+
+@dataclass(frozen=True)
+class DetectedFrame:
+    frame: int
+    occurrence: datetime
+    detections: list[Detection]
