+++ OTVision/domain/object_detection.py
@@ -1,26 +1,8 @@
 from abc import ABC, abstractmethod
-from dataclasses import dataclass
-from datetime import datetime
 from typing import Generator, Self
 
 from OTVision.config import DetectConfig
-
-
-@dataclass(frozen=True, repr=True)
-class Detection:
-    label: str
-    conf: float
-    x: float
-    y: float
-    w: float
-    h: float
-
-
-@dataclass(frozen=True)
-class DetectedFrame:
-    frame: int
-    occurrence: datetime
-    detections: list[Detection]
+from OTVision.domain.detection import Detection
 
 
 class ObjectDetector(ABC):
